XRDDEV-778 fix todos
