+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/openapi/TokenCertificatesApiController.java
@@ -31,9 +31,11 @@ import org.niis.xroad.restapi.converter.StateChangeActionConverter;
 import org.niis.xroad.restapi.converter.TokenCertificateConverter;
 import org.niis.xroad.restapi.openapi.model.StateChangeAction;
 import org.niis.xroad.restapi.openapi.model.TokenCertificate;
+import org.niis.xroad.restapi.service.ActionNotPossibleException;
 import org.niis.xroad.restapi.service.CertificateAlreadyExistsException;
 import org.niis.xroad.restapi.service.CertificateNotFoundException;
 import org.niis.xroad.restapi.service.ClientNotFoundException;
+import org.niis.xroad.restapi.service.CsrNotFoundException;
 import org.niis.xroad.restapi.service.GlobalConfService;
 import org.niis.xroad.restapi.service.KeyNotFoundException;
 import org.niis.xroad.restapi.service.StateChangeActionEnum;
@@ -84,7 +86,7 @@ public class TokenCertificatesApiController implements TokenCertificatesApi {
                 | TokenCertificateService.InvalidCertificateException
                 | TokenCertificateService.AuthCertificateNotSupportedException e) {
             throw new BadRequestException(e);
-        } catch (CertificateAlreadyExistsException | TokenCertificateService.CsrNotFoundException e) {
+        } catch (CertificateAlreadyExistsException | CsrNotFoundException e) {
             throw new ConflictException(e);
         }
         TokenCertificate tokenCertificate = tokenCertificateConverter.convert(certificate);
@@ -117,11 +119,11 @@ public class TokenCertificatesApiController implements TokenCertificatesApi {
                 | TokenCertificateService.InvalidCertificateException
                 | TokenCertificateService.AuthCertificateNotSupportedException e) {
             throw new BadRequestException(e);
-        } catch (CertificateAlreadyExistsException | TokenCertificateService.CsrNotFoundException e) {
+        } catch (CertificateAlreadyExistsException | CsrNotFoundException e) {
             throw new ConflictException(e);
         } catch (CertificateNotFoundException e) {
             throw new ResourceNotFoundException(e);
-        } catch (TokenCertificateService.ActionNotPossibleException e) {
+        } catch (ActionNotPossibleException e) {
             throw new ConflictException(e);
         }
         TokenCertificate tokenCertificate = tokenCertificateConverter.convert(certificate);
@@ -139,7 +141,7 @@ public class TokenCertificatesApiController implements TokenCertificatesApi {
         } catch (TokenCertificateService.KeyNotOperationalException
                 | TokenCertificateService.SignerOperationFailedException e) {
             throw new InternalServerErrorException(e);
-        } catch (TokenCertificateService.ActionNotPossibleException e) {
+        } catch (ActionNotPossibleException e) {
             throw new ConflictException(e);
         }
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -149,7 +151,6 @@ public class TokenCertificatesApiController implements TokenCertificatesApi {
     @PreAuthorize("hasAuthority('VIEW_KEYS')")
     public ResponseEntity<List<StateChangeAction>> getPossibleActionsForCertificate(String hash) {
         try {
-            // TO DO: test
             EnumSet<StateChangeActionEnum> actions = tokenCertificateService
                     .getPossibleActionsForCertificate(hash);
             return new ResponseEntity<>(stateChangeActionConverter.convert(actions), HttpStatus.OK);
